+++ coffeescript/templatetags/coffeescript.py
@@ -1,7 +1,8 @@
 from ..cache import get_cache_key, get_hexdigest, get_hashed_mtime
 from django.contrib.staticfiles import finders
 from ..settings import COFFEESCRIPT_EXECUTABLE, COFFEESCRIPT_USE_CACHE,\
-    COFFEESCRIPT_CACHE_TIMEOUT, COFFEESCRIPT_OUTPUT_DIR, POSIX_COMPATIBLE
+    COFFEESCRIPT_CACHE_TIMEOUT, COFFEESCRIPT_ROOT, COFFEESCRIPT_OUTPUT_DIR,\
+    POSIX_COMPATIBLE
 from django.conf import settings
 from django.core.cache import cache
 from django.template.base import Library, Node, TemplateSyntaxError
@@ -11,7 +12,12 @@ import subprocess
 import os
 
 
+STATIC_ROOT = getattr(settings, "STATIC_ROOT", getattr(settings, "MEDIA_ROOT"))
+
+
 logger = logging.getLogger("coffeescript")
+
+
 register = Library()
 
 
@@ -21,7 +27,9 @@ class InlineCoffeescriptNode(Node):
         self.nodelist = nodelist
 
     def compile(self, source):
-        args = shlex.split("%s -c -s -p" % COFFEESCRIPT_EXECUTABLE, posix=POSIX_COMPATIBLE)
+        args = shlex.split(
+            "%s -c -s -p" % COFFEESCRIPT_EXECUTABLE, posix=POSIX_COMPATIBLE
+        )
 
         p = subprocess.Popen(args, stdin=subprocess.PIPE, stdout=subprocess.PIPE)
         out, errors = p.communicate(source.encode("utf-8"))
@@ -56,11 +64,6 @@ def do_inlinecoffeescript(parser, token):
 
 def coffeescript_paths(path):
 
-    try:
-        STATIC_ROOT = settings.STATIC_ROOT
-    except AttributeError:
-        STATIC_ROOT = settings.MEDIA_ROOT
-
     full_path = os.path.join(STATIC_ROOT, path)
 
     if settings.DEBUG and not os.path.exists(full_path):
@@ -73,17 +76,17 @@ def coffeescript_paths(path):
             raise TemplateSyntaxError("Can't find staticfile named: {}".format(path))
 
     file_name = os.path.split(path)[-1]
-    output_dir = os.path.join(STATIC_ROOT, COFFEESCRIPT_OUTPUT_DIR, os.path.dirname(path))
-    
+    output_dir = os.path.join(COFFEESCRIPT_ROOT, COFFEESCRIPT_OUTPUT_DIR, os.path.dirname(path))
+
     return full_path, file_name, output_dir
 
 
 @register.simple_tag
 def coffeescript(path):
     logger.info("processing file %s" % path)
-    
+
     full_path, file_name, output_dir = coffeescript_paths(path)
-    
+
     hashed_mtime = get_hashed_mtime(full_path)
 
     base_file_name = file_name.replace(".coffee","")
@@ -96,7 +99,10 @@ def coffeescript(path):
         source = source_file.read()
         source_file.close()
 
-        args = shlex.split("%s -c -s -p" % COFFEESCRIPT_EXECUTABLE, posix=POSIX_COMPATIBLE)
+        args = shlex.split(
+            "%s -c -s -p" % COFFEESCRIPT_EXECUTABLE,
+             posix=POSIX_COMPATIBLE
+        )
         p = subprocess.Popen(args, stdin=subprocess.PIPE,
                              stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         out, errors = p.communicate(source)